Index and root page created

+++ db/seeds.rb
@@ -2,10 +2,12 @@ require 'json'
 require 'open-uri'
 
 puts "Cleaning DB..."
-Movie.destroy_all
+List.destroy_all if Rails.env.development?
+Movie.destroy_all if Rails.env.development?
+
 puts "DB cleaned"
 
-puts "Seeding..."
+puts "Seeding movies..."
 
 api_key = '10905904ad5de9b8c29f5aec179c5077'
 api_url = "https://api.themoviedb.org/3/movie/top_rated?api_key=#{api_key}"
@@ -20,6 +22,42 @@ top_movies.each do |top_movie|
     rating: top_movie["vote_average"])
 end
 
+puts "Seeding lists..."
+
+hindi_movies = List.new(name: "Hindi movies")
+hindi_movies.save!
+crime_movies = List.new(name: "Crime movies")
+crime_movies.save!
+japanese_movies = List.new(name: "Japanese movies")
+japanese_movies.save!
+
+puts "Seeding bookmarks..."
+
+movies = Movie.all.to_a
+
+rand(1..3).times do
+  Bookmark.create!(
+    comment: "#{Faker::Emotion.adjective.capitalize}, #{Faker::Emotion.adjective} and #{Faker::Emotion.adjective}!",
+    movie_id: movies.shuffle!.pop.id,
+    list_id: hindi_movies.id
+  )
+end
+
+rand(1..3).times do
+  Bookmark.create!(
+    comment: "#{Faker::Emotion.adjective.capitalize}, #{Faker::Emotion.adjective} and #{Faker::Emotion.adjective}!",
+    movie_id: movies.shuffle!.pop.id,
+    list_id: crime_movies.id
+  )
+end
+
+rand(1..3).times do
+  Bookmark.create!(
+    comment: "#{Faker::Emotion.adjective.capitalize}, #{Faker::Emotion.adjective} and #{Faker::Emotion.adjective}!",
+    movie_id: movies.shuffle!.pop.id,
+    list_id: japanese_movies.id
+  )
+end
 
 puts "Seeding completed:"
 puts "#{Movie.count} movies created"
